backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python-headless
numpy
python-dotenv
groq
reportlab
requests
pymupdf
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl